+++ src/main/resources/templates/admin/notice.html
@@ -73,10 +73,8 @@
                         </button>
                     </div>
                     <div class="modal-body">
-                        <form id="noticeAddForm" method="post"
-                              action="${pageContext.request.contextPath}/admin/notice/insert">
+                        <form id="noticeAddForm">
                             <input type="hidden" id="noticeId">
-
                             <!-- 제목과 날짜 필드를 나란히 배치 -->
                             <div class="form-group row">
                                 <!-- 제목 입력 필드 -->
