+++ src/main/resources/templates/admin/resource.html
@@ -20,8 +20,11 @@
                             <i class="fas fa-cogs"></i>
                         </div>
                         <div>
-                            <button type="button" class="btn btn-success btn-sm me-2" data-bs-toggle="modal" data-bs-target="#addResourceModal">등록</button>
-                            <button type="button" class="btn btn-secondary btn-sm me-2" onclick="openEditModal()">수정</button>
+                            <button type="button" class="btn btn-success btn-sm me-2" data-bs-toggle="modal"
+                                    data-bs-target="#addResourceModal">등록
+                            </button>
+                            <button type="button" class="btn btn-secondary btn-sm me-2" onclick="openEditModal()">수정
+                            </button>
                             <button type="button" class="btn btn-danger btn-sm" id="deleteButton">삭제</button>
                         </div>
                     </div>
